RAG-app/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
chromadb==0.5.5
pydantic==2.9.2
python-dotenv==1.0.1
pdfminer.six==20231228
python-docx==1.1.2

